fudge

+++ src/elasticnet/ConnectionGene.java
@@ -7,7 +7,7 @@ import io.netty.util.internal.ThreadLocalRandom;
 
 public class ConnectionGene implements IConnection {
 
-	public int inno_id;
+	public long inno_id;
 	int from_node = -1;
 	int to_node = -1;
 	int activation_level = 0;
